chore: Renamed the module index to validate.
