+++ lib/validate.ts
